requirements.txt
aiosignal==1.3.2
anyio==4.8.0
async-timeout==4.0.3
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-core==0.3.34
langchain-text-splitters==0.3.6
langsmith==0.3.8
numpy==1.26.4
openai==1.61.1
orjson==3.10.15
packaging==24.2
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
